app/requirements.txt
GoogleAppEngineCloudStorageClient==1.9.22.1
Pillow==5.3.0
